reworking the dialogs into BPI
